pyproject.toml
[tool.poetry]
name = "spending-tracker"
version = "0.1.0"
description = "We pray to our lord and savior that this works"
authors = ["Yavor"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.10.*"
matplotlib = "^3.8.3"
flask = "^3.0.3"
psycopg2 = "^2.9.9"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.1"
pytest = "^8.3.2"
pytest-flask = "^1.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
